Added implementation for wiringPi driver
In case the wiringPi and the RF24 are used together in the same
application CPU load goes to 100%. The solution was to create a new
implementation based on the SPI from wiringPi library. In order to
have it working the wiringPi library should be already installed.
For enabling this implementation the "--driver=wiringPi" should be
considered when compiling the RF24.

Fixes the reported issues from #305 ticket.

Signed-off-by: Aurelian Z. <aurelian17@gmail.com>